+++ src/main/java/unah/hn/lenguajes1900/info/demo/repositories/VehiculoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import unah.hn.lenguajes1900.info.demo.entities.Vehiculo;
 
-public interface VehiculoRepository extends CrudRepository<Vehiculo, Integer>{
+public interface VehiculoRepository extends CrudRepository<Vehiculo, Long>{
     
 }
