+++ src/main/java/unah/hn/lenguajes1900/info/demo/services/VehiculoService.java
@@ -10,5 +10,5 @@ public interface VehiculoService {
 
     public List<Vehiculo> obtenerVehiculos();
 
-    public Vehiculo obtenerVehiculoPorId(Integer idVehiculo );
+    public Vehiculo obtenerVehiculoPorId(Long idVehiculo );
 }
